// src/app/enums/page-name.enum.ts
export enum PageName {
    Welcome = 'welcome',
    Login = 'login',
    ExistingIra = 'existing-ira',
    Situations = 'situations',
    AskDob = 'ask-dob',
    TaxFiling = 'tax-filing',
    GrossIncome = 'gross-income',
    AccountChoices = 'account-choices',
    AreYouClient = 'are-you-client',
    OpenOrEnroll = 'open-or-enroll',
    NewAccountTypes = 'new-account-types',
    ExistingAccounts = 'existing-accounts'
}